APIO-103 Reorder wiring-osgi packages
