+++ modules/apps/foundation/apio-architect/apio-architect-application/src/main/java/com/liferay/apio/architect/application/internal/endpoint/RootEndpointImpl.java
@@ -35,12 +35,12 @@ import com.liferay.apio.architect.routes.NestedCollectionRoutes;
 import com.liferay.apio.architect.single.model.SingleModel;
 import com.liferay.apio.architect.uri.Path;
 import com.liferay.apio.architect.url.ServerURL;
-import com.liferay.apio.architect.wiring.osgi.manager.CollectionRouterManager;
-import com.liferay.apio.architect.wiring.osgi.manager.ItemRouterManager;
-import com.liferay.apio.architect.wiring.osgi.manager.NestedCollectionRouterManager;
 import com.liferay.apio.architect.wiring.osgi.manager.ProviderManager;
-import com.liferay.apio.architect.wiring.osgi.manager.RepresentableManager;
-import com.liferay.apio.architect.wiring.osgi.manager.ReusableNestedCollectionRouterManager;
+import com.liferay.apio.architect.wiring.osgi.manager.representable.RepresentableManager;
+import com.liferay.apio.architect.wiring.osgi.manager.router.CollectionRouterManager;
+import com.liferay.apio.architect.wiring.osgi.manager.router.ItemRouterManager;
+import com.liferay.apio.architect.wiring.osgi.manager.router.NestedCollectionRouterManager;
+import com.liferay.apio.architect.wiring.osgi.manager.router.ReusableNestedCollectionRouterManager;
 
 import java.io.InputStream;
