+++ modules/apps/foundation/apio-architect/apio-architect-jaxrs-json/src/main/java/com/liferay/apio/architect/jaxrs/json/internal/writer/SingleModelMessageBodyWriter.java
@@ -28,7 +28,7 @@ import com.liferay.apio.architect.single.model.SingleModel;
 import com.liferay.apio.architect.url.ServerURL;
 import com.liferay.apio.architect.wiring.osgi.manager.PathIdentifierMapperManager;
 import com.liferay.apio.architect.wiring.osgi.manager.ProviderManager;
-import com.liferay.apio.architect.wiring.osgi.manager.RepresentableManager;
+import com.liferay.apio.architect.wiring.osgi.manager.representable.RepresentableManager;
 import com.liferay.apio.architect.wiring.osgi.util.GenericUtil;
 import com.liferay.apio.architect.writer.SingleModelWriter;
