+++ modules/apps/foundation/apio-architect/apio-architect-wiring-osgi-api/src/main/java/com/liferay/apio/architect/wiring/osgi/manager/router/NestedCollectionRouterManager.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.apio.architect.wiring.osgi.manager;
+package com.liferay.apio.architect.wiring.osgi.manager.router;
 
 import aQute.bnd.annotation.ProviderType;
