+++ modules/apps/foundation/apio-architect/apio-architect-wiring-osgi-impl/src/main/java/com/liferay/apio/architect/wiring/osgi/internal/manager/router/CollectionRouterManagerImpl.java
@@ -12,27 +12,29 @@
  * details.
  */
 
-package com.liferay.apio.architect.wiring.osgi.internal.manager;
-
-import static com.liferay.apio.architect.wiring.osgi.internal.manager.ManagerUtil.getTypeParamOrFail;
+package com.liferay.apio.architect.wiring.osgi.internal.manager.router;
 
 import static org.osgi.service.component.annotations.ReferenceCardinality.MULTIPLE;
 import static org.osgi.service.component.annotations.ReferencePolicy.DYNAMIC;
 import static org.osgi.service.component.annotations.ReferencePolicyOption.GREEDY;
 
 import com.liferay.apio.architect.alias.RequestFunction;
-import com.liferay.apio.architect.identifier.Identifier;
-import com.liferay.apio.architect.router.ReusableNestedCollectionRouter;
-import com.liferay.apio.architect.routes.NestedCollectionRoutes;
-import com.liferay.apio.architect.routes.NestedCollectionRoutes.Builder;
+import com.liferay.apio.architect.router.CollectionRouter;
+import com.liferay.apio.architect.routes.CollectionRoutes;
+import com.liferay.apio.architect.routes.CollectionRoutes.Builder;
+import com.liferay.apio.architect.wiring.osgi.internal.manager.base.BaseManager;
 import com.liferay.apio.architect.wiring.osgi.manager.ProviderManager;
-import com.liferay.apio.architect.wiring.osgi.manager.RepresentableManager;
-import com.liferay.apio.architect.wiring.osgi.manager.ReusableNestedCollectionRouterManager;
+import com.liferay.apio.architect.wiring.osgi.manager.representable.RepresentableManager;
+import com.liferay.apio.architect.wiring.osgi.manager.router.CollectionRouterManager;
 
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.Function;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import org.osgi.framework.ServiceReference;
 import org.osgi.service.component.annotations.Component;
@@ -42,14 +44,13 @@ import org.osgi.service.component.annotations.Reference;
  * @author Alejandro Hernández
  */
 @Component(immediate = true)
-public class ReusableNestedCollectionRouterManagerImpl
-	extends BaseManager<ReusableNestedCollectionRouter>
-	implements ReusableNestedCollectionRouterManager {
+public class CollectionRouterManagerImpl
+	extends BaseManager<CollectionRouter> implements CollectionRouterManager {
 
 	@Override
 	@SuppressWarnings("unchecked")
-	public <T> Optional<NestedCollectionRoutes<T>>
-		getNestedCollectionRoutesOptional(String name) {
+	public <T> Optional<CollectionRoutes<T>> getCollectionRoutesOptional(
+		String name) {
 
 		Optional<Class<T>> optional =
 			_representableManager.getModelClassOptional(name);
@@ -57,42 +58,58 @@ public class ReusableNestedCollectionRouterManagerImpl
 		return optional.map(
 			Class::getName
 		).map(
-			_routesMap::get
+			_collectionRoutesMap::get
 		).map(
-			routes -> (NestedCollectionRoutes<T>)routes
+			routes -> (CollectionRoutes<T>)routes
+		);
+	}
+
+	@Override
+	public List<String> getResourceNames() {
+		Set<String> keys = _collectionRoutesMap.keySet();
+
+		Stream<String> stream = keys.stream();
+
+		return stream.map(
+			className -> _representableManager.getNameOptional(className)
+		).filter(
+			Optional::isPresent
+		).map(
+			Optional::get
+		).collect(
+			Collectors.toList()
 		);
 	}
 
 	@Reference(cardinality = MULTIPLE, policy = DYNAMIC, policyOption = GREEDY)
 	protected void setServiceReference(
-		ServiceReference<ReusableNestedCollectionRouter> serviceReference) {
+		ServiceReference<CollectionRouter> serviceReference) {
 
 		Optional<Class<Object>> optional = addService(
-			serviceReference, ReusableNestedCollectionRouter.class);
+			serviceReference, CollectionRouter.class);
 
 		optional.ifPresent(this::_addRoutes);
 	}
 
 	@SuppressWarnings("unused")
 	protected void unsetServiceReference(
-		ServiceReference<ReusableNestedCollectionRouter> serviceReference) {
+		ServiceReference<CollectionRouter> serviceReference) {
 
 		Optional<Class<Object>> optional = removeService(
-			serviceReference, ReusableNestedCollectionRouter.class);
+			serviceReference, CollectionRouter.class);
 
 		optional.map(
 			Class::getName
 		).ifPresent(
-			_routesMap::remove
+			_collectionRoutesMap::remove
 		);
 
 		optional.filter(
 			modelClass -> {
-				Optional<ReusableNestedCollectionRouter>
-					nestedCollectionRouterOptional = getServiceOptional(
-						modelClass);
+				Optional<CollectionRouter> collectionRouterOptional =
+					getServiceOptional(modelClass);
 
-				return nestedCollectionRouterOptional.isPresent();
+				return collectionRouterOptional.isPresent();
 			}
 		).ifPresent(
 			this::_addRoutes
@@ -100,44 +117,38 @@ public class ReusableNestedCollectionRouterManagerImpl
 	}
 
 	@SuppressWarnings("unchecked")
-	private <T, U extends Identifier> void _addRoutes(Class<T> modelClass) {
-		Optional<ReusableNestedCollectionRouter> optional = getServiceOptional(
-			modelClass);
+	private <T> void _addRoutes(Class<T> modelClass) {
+		Optional<CollectionRouter> optional = getServiceOptional(modelClass);
 
 		optional.map(
-			reusableNestedCollectionRouter ->
-				(ReusableNestedCollectionRouter<T, U>)
-					reusableNestedCollectionRouter
+			collectionRouter -> (CollectionRouter<T>)collectionRouter
 		).ifPresent(
-			reusableNestedCollectionRouter -> {
-				Class<U> identifierClass = getTypeParamOrFail(
-					reusableNestedCollectionRouter,
-					ReusableNestedCollectionRouter.class, 1);
-
+			collectionRouter -> {
 				RequestFunction<Function<Class<?>, Optional<?>>>
 					provideClassFunction =
 						httpServletRequest -> clazz ->
 							_providerManager.provideOptional(
 								clazz, httpServletRequest);
 
-				Builder<T, U> builder = new Builder<>(
-					modelClass, identifierClass, provideClassFunction);
+				Builder<T> builder = new Builder<>(
+					modelClass, provideClassFunction);
 
-				NestedCollectionRoutes<T> routes =
-					reusableNestedCollectionRouter.collectionRoutes(builder);
+				CollectionRoutes<T> collectionRoutes =
+					collectionRouter.collectionRoutes(builder);
 
-				_routesMap.put(modelClass.getName(), routes);
+				_collectionRoutesMap.put(
+					modelClass.getName(), collectionRoutes);
 			}
 		);
 	}
+
+	private final Map<String, CollectionRoutes<?>> _collectionRoutesMap =
+		new ConcurrentHashMap<>();
 
 	@Reference
 	private ProviderManager _providerManager;
 
 	@Reference
 	private RepresentableManager _representableManager;
-
-	private final Map<String, NestedCollectionRoutes<?>> _routesMap =
-		new ConcurrentHashMap<>();
 
 }
